==> urbandata/01-cleaning.ipynb <==
#this notebook is use to clean the collected data

import pandas as pd
import re

# Quinto Andar - clean rent apartament raw-data

## load

df_qa_apartment_raw = pd.read_csv('../data/raw-data/df_qa_apartment_raw.csv')

df_qa_apartment_raw.head(2)

def split_elements(row):
    row = re.sub('\'|\[|\]|"','',row).split(',')
    return row

df_qa_apartment_raw['endereco'] = df_qa_apartment_raw['endereco'].apply(split_elements)

addresses = df_qa_apartment_raw.endereco.apply(lambda x : pd.Series(x))

addresses

## functions

def find_numero(row):
    numero = re.findall('\d+',row)
    return numero

def find_cep(row):
    return re.findall('\d+-\d+', row)

def find_lat_long(row):
    if type(row) == str:
        return re.findall('\d+.\d+',row)
    else:
        pass

def acessa_lista(row):
    if type(row) == list:
        for x in row:
            return x
    else:
        return row  

def clean_metragem_quartos(row):
    numero = re.findall('\d+', row)
    return numero

def clean_precos(row):
    if type(row) == str:
        row = row.replace('.','')
        row = row.replace(',','')
        valor = re.findall('\d+', row)
        return valor
    else:
        return row

def acessa_lista_int(row):
    if type(row) == list:
        for x in row:
            return int(x)
    else:
        return row  

def negative_numbers(row):
    if type(row) == str:
        row = float(row)*-1
        return row
    else:
        return row

lista_tipologias = ['apartamento', 'studio e kitnet','casa']

def tipologia(row, lista_tipologias=lista_tipologias):
    for tipo in lista_tipologias:
        if re.findall(tipo, row, flags=re.IGNORECASE):
            return tipo
        
    return 'outro'

## cleaning data

addresses['longitude'] = addresses.iloc[:,[1]]
addresses['latitude'] = addresses.iloc[:,[9]]

addresses['numero'] = addresses.iloc[:,[2]]
addresses['cep'] = addresses.iloc[:,[4]]
addresses = addresses[['numero','cep','longitude','latitude']]

addresses['numero'] = addresses['numero'].apply(find_numero)
addresses['cep'] = addresses['cep'].apply(find_cep)
addresses['longitude'] = addresses['longitude'].apply(find_lat_long)
addresses['latitude'] = addresses['latitude'].apply(find_lat_long)

addresses['numero'] = addresses['numero'].apply(acessa_lista)
addresses['cep'] = addresses['cep'].apply(acessa_lista)
addresses['longitude'] = addresses['longitude'].apply(acessa_lista)
addresses['latitude'] = addresses['latitude'].apply(acessa_lista)

df_qa_apartment = pd.concat([df_qa_apartment_raw, addresses], axis=1)
df_qa_apartment = df_qa_apartment.drop(columns='endereco')

df_qa_apartment.head()

df_qa_apartment['tipo_imovel'] = df_qa_apartment['tipo_imovel'].apply(tipologia)

df_qa_apartment['metragem'] = df_qa_apartment['metragem'].apply(clean_metragem_quartos)
df_qa_apartment['quartos'] = df_qa_apartment['quartos'].apply(clean_metragem_quartos)
df_qa_apartment['banheiros'] = df_qa_apartment['banheiros'].apply(clean_metragem_quartos)

df_qa_apartment['vaga'] = df_qa_apartment['vaga'].apply(clean_metragem_quartos)
df_qa_apartment['andar'] = df_qa_apartment['andar'].apply(clean_metragem_quartos)
df_qa_apartment['aluguel'] = df_qa_apartment['aluguel'].apply(clean_precos)
df_qa_apartment['condominio'] = df_qa_apartment['condominio'].apply(clean_precos)
df_qa_apartment['iptu'] = df_qa_apartment['iptu'].apply(clean_precos)
df_qa_apartment['preco_total'] = df_qa_apartment['preco_total'].apply(clean_precos)

df_qa_apartment['longitude'] = df_qa_apartment['longitude'].apply(negative_numbers)
df_qa_apartment['latitude'] = df_qa_apartment['latitude'].apply(negative_numbers)

df_qa_apartment['aluguel'] = df_qa_apartment['aluguel'].apply(acessa_lista)
df_qa_apartment['condominio'] = df_qa_apartment['condominio'].apply(acessa_lista)
df_qa_apartment['iptu'] = df_qa_apartment['iptu'].apply(acessa_lista)
df_qa_apartment['preco_total'] = df_qa_apartment['preco_total'].apply(acessa_lista)
df_qa_apartment['metragem'] = df_qa_apartment['metragem'].apply(acessa_lista)
df_qa_apartment['quartos'] = df_qa_apartment['quartos'].apply(acessa_lista)
df_qa_apartment['banheiros'] = df_qa_apartment['banheiros'].apply(acessa_lista)
df_qa_apartment['vaga'] = df_qa_apartment['vaga'].apply(acessa_lista)
df_qa_apartment['andar'] = df_qa_apartment['andar'].apply(acessa_lista)

df_qa_apartment.head()

## latitude and longitude to utm 

dict_utm = {}

for row in range(df_quinto_andar.shape[0]):

    lat = df_quinto_andar['latitude'][row]
    long = df_quinto_andar['longitude'][row]
    try:
        y, x, z, k = utm.from_latlon(lat, long)
        dict_row_utm = {}
        dict_row_utm = {'latitude_utm': round(x,2) , 'longitude_utm' : round(y,2)}
        dict_utm[row] = dict_row_utm
    except:
        dict_utm[row] = dict_row_utm
        dict_row_utm = {'latitude_utm': 0 , 'longitude_utm' : 0}

df_quinto_andar_utm = pd.DataFrame.from_dict(dict_utm, orient='index')
df_qa_clean =pd.concat([df_quinto_andar, df_quinto_andar_utm], axis=1)

## save

df_qa_clean.to_csv('../data/clean-data/df_qa_clean.csv', index = False)

# Loft - apartment for sale

## load 

df_loft_resumo = df_loft_resumo.reset_index(drop=True)

## latitude and longitude to utm 

dict_utm = {}

for row in range(df_loft_resumo.shape[0]):

    lat = df_loft_resumo['latitude'][row]
    long = df_loft_resumo['longitude'][row]
    try:
        y, x, z, k = utm.from_latlon(lat, long)
        dict_row_utm = {}
        dict_row_utm = {'latitude_utm': round(x,2) , 'longitude_utm' : round(y,2)}
        dict_utm[row] = dict_row_utm
    except:
        dict_utm[row] = dict_row_utm
        dict_row_utm = {'latitude_utm': 0 , 'longitude_utm' : 0}

df_loft_resumo_utm = pd.DataFrame.from_dict(dict_utm, orient='index')
df_loft_resumo =pd.concat([df_loft_resumo, df_loft_resumo_utm], axis=1)

mask = (df_loft_resumo['latitude_utm'] > 0 ) & (df_loft_resumo['longitude_utm'] > 0 )
df_loft_resumo = df_loft_resumo.loc[mask,:]

## save 

df_loft_clean.to_csv('../data/clean-data/clean-data/df_loft_clean.csv', index = False)

# Google - places

## load 

df_google_raw = pd.read_csv('../data/raw-data/df_google_raw.csv')

## functions

def retorna_tipo_lugar(row):
    if ('bakery' in row) or ('bar' in row) or ('cafe' in row) or ('restaurant' in row):
        return 'lugares_alimentacao'
    
    elif ('bank' in row) or ('beauty_salon' in row) or ('dentist' in row) or ('gym' in row) or ('laundry' in row) or ('lawyer' in row):
        return 'lugares_servicos'
    
    elif ('shopping_mall' in row) or ('supermarket' in row) or ('drugstore' in row) or ('florist' in row) or ('gas_station' in row) or ('parking' in row) or ('pharmacy' in row):
        return 'lugares_comercio'
    
    elif ('subway_station' in row) or ('train_station' in row):
        return 'lugares_metro_trem'
    
    elif ('doctor' in row) or ('hospital' in row):
        return 'lugares_saude'
    
    elif ('library' in row) or ('primary_school' in row) or ('school' in row) or ('secondary_school' in row) or ('university' in row):
        return 'lugares_educacao'
    
    elif ('museum' in row):
        return 'lugares_cultura'
    
    elif ('church' in row) or ('cemetery' in row) or ('synagogue' in row):
        return 'lugares_religiosos'
    
    elif ('park' in row):
        return 'lugares_parques'
    
    elif ('fire_station' in row) or ('police' in row) or ('post_office' in row):
        return 'lugares_governamentais'
    
    elif ('stadium' in row):
        return 'lugares_estadios'
    
    elif ('bus_station' in row):
        return 'lugares_transporte'
    
    elif ('tourist_attraction' in row):
        return 'lugares_turismo'
    
    else:
        return 'outros'

## cleaning

df_google = df_google.drop_duplicates()

localizacao = df_google.geometry.apply(lambda x : pd.Series(eval(x)))
localizacao = localizacao['location'].apply(lambda x : pd.Series(x))
df_google_final = pd.concat([df_google, localizacao], axis=1)

df_google_final.columns

df_google_final = df_google_final.drop(columns=['business_status','geometry','icon','opening_hours','photos','reference','permanently_closed','place_id','plus_code'])

df_google_final['lugares']  = df_google_final['types'].apply(retorna_tipo_lugar)

df_google_final = df_google_final.reset_index(drop=True)

df_google_final['lugares'].unique()

## latitude and longitude to utm 

dict_utm = {}

for row in range(df_google_final.shape[0]):

    lat = df_google_final['lat'][row]
    long = df_google_final['lng'][row]

    y, x, z, k = utm.from_latlon(lat, long)
    dict_row_utm = {}
    dict_row_utm = {'latitude_utm': round(x,2) , 'longitude_utm' : round(y,2)}
    dict_utm[row] = dict_row_utm

df_google_utm = pd.DataFrame.from_dict(dict_utm, orient='index')
df_google_final =pd.concat([df_google_final, df_google_utm], axis=1)

## save 

df_google_clean.to_csv('../data/clean-data/df_google_clean.csv', index=False)
