==> urbandata/02-merge.ipynb <==
import pandas as pd
from collections import Counter
import matplotlib.pyplot as plt
import numpy as np
import seaborn as sns

df_loft_clean = pd.read_csv('../data/clean-data/df_loft_clean.csv')
df_qa_clean = pd.read_csv('../data/clean-data/df_qa_clean.csv')
df_google_clean = pd.read_csv('../data/clean-data/df_google_clean.csv')

# Merge

## Quinto Andar - Google

dict_places_near = {}

for row_quinto in range(df_quinto_andar.shape[0]):
    #print('row quinto ' ,row_quinto )
    latitude_utm_ape = df_quinto_andar['latitude_utm'][row_quinto]
    longitude_utm_ape = df_quinto_andar['longitude_utm'][row_quinto]
    
    #print('latitude e longitude' ,latitude_utm_ape ,longitude_utm_ape F)
    
    list_categorias = [] 

    for row_empresa in range(df_google_final.shape[0]):
        #print(row_empresa)
        
        latitude_empresa = df_google_final['latitude_utm'][row_empresa]
        longitude_empresa = df_google_final['longitude_utm'][row_empresa]
        #print('latitude e longitude empresa', latitude_empresa , longitude_empresa)
        

        distancia = 1000
        condition1 = latitude_utm_ape - distancia < latitude_empresa < latitude_utm_ape + distancia
        condition2 = longitude_utm_ape - distancia < longitude_empresa < longitude_utm_ape + distancia
    
        if condition1 and condition2:
            list_categorias.append(df_google_final['lugares'][row_empresa])
        total_empresas = Counter(list_categorias)
        #print(total_empresas)
    dict_places_near[row_quinto] = total_empresas

dict_places_near

df_catedorias = pd.DataFrame.from_dict(dict_places_near, orient='index').fillna(0)


df_quinto_empresas = pd.concat([df_quinto_andar, df_catedorias], axis=1).fillna(0)

df_qa_merge = df_quinto_empresas

df_qa_merge.to_csv('data/merge-data/df_qa_merge.csv', index=False)

## Loft - Google 

dict_places_near = {}

for row_loft in range(df_loft.shape[0]):
    
    latitude_utm_ape = df_loft['latitude_utm'][row_loft]
    longitude_utm_ape = df_loft['longitude_utm'][row_loft]
    
    #print('latitude e longitude' ,latitude_utm_ape ,longitude_utm_ape F)
    
    list_categorias = [] 

    for row_empresa in range(df_google_final2.shape[0]):
        #print(row_empresa)
        
        latitude_empresa = df_google_final2['latitude_utm'][row_empresa]
        longitude_empresa = df_google_final2['longitude_utm'][row_empresa]
        #print('latitude e longitude empresa', latitude_empresa , longitude_empresa)
        

        distancia = 1000
        condition1 = latitude_utm_ape - distancia < latitude_empresa < latitude_utm_ape + distancia
        condition2 = longitude_utm_ape - distancia < longitude_empresa < longitude_utm_ape + distancia
    
        if condition1 and condition2:
            list_categorias.append(df_google_final2['lugares'][row_empresa])
        total_empresas = Counter(list_categorias)
        #print(total_empresas)
    dict_places_near[row_loft] = total_empresas

df_catedorias = pd.DataFrame.from_dict(dict_places_near, orient='index').fillna(0)
df_loft_google = pd.concat([df_loft, df_catedorias], axis=1).fillna(0)

df_loft_google.head()

df_loft_merge.to_csv('data/merge-data/df_loft_merge.csv', index=False)
